Removes base class for brevity
