+++ Common/Data/Consolidators/RenkoConsolidator.cs
@@ -22,8 +22,41 @@ namespace QuantConnect.Data.Consolidators
     /// <summary>
     /// This consolidator can transform a stream of <see cref="BaseData"/> instances into a stream of <see cref="RenkoBar"/>
     /// </summary>
-    public class RenkoConsolidator : BaseRenkoConsolidator
+    public class RenkoConsolidator : IDataConsolidator
     {
+        private RenkoBar _currentBar;
+        private DataConsolidatedHandler _dataConsolidatedHandler;
+        private decimal _barSize;
+        private bool _evenBars;
+        private Func<IBaseData, decimal> _selector;
+        private Func<IBaseData, decimal> _volumeSelector;
+
+        /// <summary>
+        /// Gets the kind of the bar
+        /// </summary>
+        public RenkoType Type => RenkoType.Classic;
+
+        /// <summary>
+        /// Gets a clone of the data being currently consolidated
+        /// </summary>
+        public IBaseData WorkingData => _currentBar?.Clone();
+
+        /// <summary>
+        /// Gets the type consumed by this consolidator
+        /// </summary>
+        public Type InputType => typeof(IBaseData);
+
+        /// <summary>
+        /// Gets <see cref="RenkoBar"/> which is the type emitted in the <see cref="IDataConsolidator.DataConsolidated"/> event.
+        /// </summary>
+        public Type OutputType => typeof(RenkoBar);
+
+        /// <summary>
+        /// Gets the most recently consolidated piece of data. This will be null if this consolidator
+        /// has not produced any data yet.
+        /// </summary>
+        public IBaseData Consolidated { get; private set; }
+
         /// <summary>
         /// Initializes a new instance of the <see cref="RenkoConsolidator"/> class using the specified <paramref name="barSize"/>.
         /// The value selector will by default select <see cref="IBaseData.Value"/>
@@ -32,8 +65,12 @@ namespace QuantConnect.Data.Consolidators
         /// <param name="barSize">The constant value size of each bar</param>
         /// <param name="evenBars">When true bar open/close will be a multiple of the barSize</param>
         public RenkoConsolidator(decimal barSize, bool evenBars = true)
-            : base(barSize, evenBars)
         {
+            EpsilonCheck(barSize);
+            _barSize = barSize;
+            _selector = x => x.Value;
+            _evenBars = evenBars;
+            _volumeSelector = x => 0;
         }
 
         /// <summary>
@@ -51,8 +88,12 @@ namespace QuantConnect.Data.Consolidators
             Func<IBaseData, decimal> volumeSelector = null,
             bool evenBars = true
             )
-            : base(barSize, selector, volumeSelector, evenBars)
         {
+            EpsilonCheck(barSize);
+            _barSize = barSize;
+            _selector = selector;
+            _evenBars = evenBars;
+            _volumeSelector = volumeSelector;
         }
 
         /// <summary>
@@ -70,45 +111,126 @@ namespace QuantConnect.Data.Consolidators
             PyObject volumeSelector = null,
             bool evenBars = true
             )
-            : base(barSize, selector, volumeSelector, evenBars)
         {
+            EpsilonCheck(barSize);
+
+            if (selector != null)
+            {
+                if (!selector.TryConvertToDelegate(out _selector))
+                {
+                    throw new ArgumentException(
+                        "Unable to convert parameter 'selector' to delegate type Func<IBaseData, decimal>");
+                }
+            }
+            else
+            {
+                _selector = x => x.Value;
+            }
+
+            if (volumeSelector != null)
+            {
+                if (!volumeSelector.TryConvertToDelegate(out _volumeSelector))
+                {
+                    throw new ArgumentException(
+                        "Unable to convert parameter 'volumeSelector' to delegate type Func<IBaseData, decimal>");
+                }
+            }
+            else
+            {
+                _volumeSelector = x => 0;
+            }
+
+            _evenBars = evenBars;
+        }
+
+        /// <summary>
+        /// Event handler that fires when a new piece of data is produced
+        /// </summary>
+        public event EventHandler<RenkoBar> DataConsolidated;
+
+        /// <summary>
+        /// Event handler that fires when a new piece of data is produced
+        /// </summary>
+        event DataConsolidatedHandler IDataConsolidator.DataConsolidated
+        {
+            add { _dataConsolidatedHandler += value; }
+            remove { _dataConsolidatedHandler -= value; }
         }
 
         /// <summary>
         /// Updates this consolidator with the specified data
         /// </summary>
         /// <param name="data">The new data for the consolidator</param>
-        public override void Update(IBaseData data)
+        public void Update(IBaseData data)
         {
-            var currentValue = Selector(data);
-            var volume = VolumeSelector(data);
+            var currentValue = _selector(data);
+            var volume = _volumeSelector(data);
 
             decimal? close = null;
 
             // if we're already in a bar then update it
-            if (CurrentBar != null)
+            if (_currentBar != null)
             {
-                CurrentBar.Update(data.Time, currentValue, volume);
+                _currentBar.Update(data.Time, currentValue, volume);
 
                 // if the update caused this bar to close, fire the event and reset the bar
-                if (CurrentBar.IsClosed)
+                if (_currentBar.IsClosed)
                 {
-                    close = CurrentBar.Close;
-                    OnDataConsolidated(CurrentBar);
-                    CurrentBar = null;
+                    close = _currentBar.Close;
+                    OnDataConsolidated(_currentBar);
+                    _currentBar = null;
                 }
             }
 
-            if (CurrentBar == null)
+            if (_currentBar == null)
             {
                 var open = close ?? currentValue;
-                if (EvenBars && !close.HasValue)
+                if (_evenBars && !close.HasValue)
                 {
-                    open = Math.Ceiling(open / BarSize) * BarSize;
+                    open = Math.Ceiling(open / _barSize) * _barSize;
                 }
 
-                CurrentBar = new RenkoBar(data.Symbol, data.Time, BarSize, open, volume);
+                _currentBar = new RenkoBar(data.Symbol, data.Time, _barSize, open, volume);
             }
+        }
+
+        /// <summary>
+        /// Scans this consolidator to see if it should emit a bar due to time passing
+        /// </summary>
+        /// <param name="currentLocalTime">The current time in the local time zone (same as <see cref="BaseData.Time"/>)</param>
+        public void Scan(DateTime currentLocalTime)
+        {
+        }
+
+        /// <summary>Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.</summary>
+        /// <filterpriority>2</filterpriority>
+        public void Dispose()
+        {
+            DataConsolidated = null;
+            _dataConsolidatedHandler = null;
+        }
+
+        private static void EpsilonCheck(decimal barSize)
+        {
+            if (barSize < Extensions.GetDecimalEpsilon())
+            {
+                throw new ArgumentOutOfRangeException(nameof(barSize),
+                    "RenkoConsolidator bar size must be positve and greater than 1e-28");
+            }
+        }
+
+        /// <summary>
+        /// Event invocator for the DataConsolidated event. This should be invoked
+        /// by derived classes when they have consolidated a new piece of data.
+        /// </summary>
+        /// <param name="consolidated">The newly consolidated data</param>
+        private void OnDataConsolidated(RenkoBar consolidated)
+        {
+            DataConsolidated?.Invoke(this, consolidated);
+
+            _dataConsolidatedHandler?.Invoke(this, consolidated);
+
+            Consolidated = consolidated;
         }
     }
 
@@ -158,7 +280,7 @@ namespace QuantConnect.Data.Consolidators
         /// Type safe shim method.
         /// </remarks>
         /// <param name="data">The new data for the consolidator</param>
-        public void Update(IBaseData data)
+        public void Update(TInput data)
         {
             base.Update(data);
         }
